Colors, Columns & Throbbing
Made some color variables for this project. There are two main color
schemes which work in the dark and light themes of the site.

Gave some grid variation. single-column, double-column, and a single
that floods between them (will all be single-column on mobile).

Added some placeholder animation via Animate.css for some of the page
hints: calendar and group of people. The group of people was a fun idea
that just came to me on the fly. Shows their solidarity and
scatterness. Thinking about how to tell a story with some more advanced
animations.